Fix: 코드리뷰 반영 (오타 수정, 타입변경)

+++ src/main/java/hana/teamfour/addminhana/entity/ProductEntity.java
@@ -15,13 +15,13 @@ public class ProductEntity {
     private String p_category;
     private String p_name;
     private String p_description;
-    private Integer p_interestrate;
+    private Double p_interestrate;
     private Integer p_intrest_day;
     private Timestamp p_date;
     private Integer p_contract_month;
     private Character p_isactive;
     private Integer p_limit;
-    private Integer p_collateralrate;
+    private Double p_collateralrate;
     private Integer p_mincreditgrade;
     private String p_jobtype;
 }
